Update various views and add in functionality
Fix table issue with organizations displaying events but not being able
to select or move table. Added in segue to OpportunitiesViewController
from OrganizationViewController when a user selects a specific event.
Created a registration page for user to register for a particular
event. Added in an organization photo to fill up white space within
OrganizationViewController. Also updated JSON to provide more realistic
descriptions and events to test various design functionalities.